test(mobile): split agent-push-preference tests under the max-lines cap
